tests: put fuzz_journald_processing_function in a .c file
(cherry picked from commit 231dca5579cfba6175d19eee5347d693893fb5aa)

Resolves: #1764560
